4.10：21提交，更新了数据库模型

+++ src/FileTool/FileTool.go
@@ -6,7 +6,21 @@ import (
 	"path"
 )
 
+func PathExists(path string) (bool, error) {
+	_, err := os.Stat(path)
+	if err == nil {
+		return true, nil
+	}
+	if os.IsNotExist(err) {
+		return false, nil
+	}
+	return false, err
+}
+
 func Clear(basePath string) error {
+	if ok, _ := PathExists(basePath); !ok {
+		return os.Mkdir(basePath, 0777)
+	}
 	err := os.RemoveAll(basePath)
 	if err != nil {
 		return err
@@ -39,4 +53,16 @@ func ReadCases(basePath string) ([]string, error) {
 
 func WriteFile(path string, str *string) error {
 	return ioutil.WriteFile(path, []byte(*str), 0777)
+}
+func WriteFileB(path string, buf []byte) error {
+	return ioutil.WriteFile(path, buf, 0777)
+}
+
+func ReadFile(path string) ([]byte, error) {
+	return ioutil.ReadFile(path)
+}
+
+func Home() string {
+	return "/code/TMManager/"
+	//return os.Getenv("CUGOJ_HOME")
 }
